requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2022.4
sqlparse==0.4.3
whitenoise==6.2.0
